feat: Add possibility to substitute DiskTrieAccessor in RuntimeImpl per method call. Refactor runtime handling when runtime code changes after executing a block. Improve overall stability.

+++ src/main/java/com/limechain/runtime/version/scale/RuntimeVersionWriter.java
@@ -37,6 +37,6 @@ public class RuntimeVersionWriter implements ScaleWriter<RuntimeVersion> {
         }
 
         // Write the state version if it's present. Older runtimes miss this field, so StateVersion 0 is to be presumed.
-        scaleCodecWriter.writeByte(runtimeVersion.getStateVersion().asInt());
+        scaleCodecWriter.writeUint32(runtimeVersion.getStateVersion().asInt());
     }
 }
